requirements.txt
argparse==1.4.0
Pillow==10.4.0
numpy==2.0.2